refactor(web): extract shared QuickKeys from the terminal view
The standalone terminal's quick-input keys (Esc/arrows/Ctrl/Alt) and their
sticky-modifier dispatch were inline in terminal.tsx. Extract them to a shared
QuickKeys module — a useQuickKeyInput hook (modifier state + dispatcher) and a
QuickKeyRows presentational component — so other terminal views can reuse them
without duplication. Behavior-preserving: the raw onData path stays ungated and
shares the modifier state with the buttons, exactly as before.
